+++ include/hm_set.h
@@ -17,7 +17,7 @@
  * The return signal of function in set
  */
 typedef enum hm_set_ret {
-    hm_set_ret_error = 0x4,     // Malloc failed
+    hm_set_ret_error = 0x16,     // Malloc failed
     hm_set_ret_warn,            // The pass parameter is incorrect
     hm_set_ret_none,            // Operation invalid, like del entry by key , but key is not existed in set
     hm_set_ret_existed,         // Insert entry in set, but key is existed , you should free the key if the key that you given is alloced